requirements.txt
mysql-connector-python
boto3
python-interface